add: new Meme endpoint

+++ src/main/java/dev/chipichapa/memestore/dto/meme/GetMemeRequest.java
@@ -1,4 +1,4 @@
 package dev.chipichapa.memestore.dto.meme;
 
-public record GetMemeRequest(int galleryId, int memeId) {
+public record GetMemeRequest(long galleryId, long memeId) {
 }
